Use integer for course.

+++ mls.h
@@ -42,7 +42,7 @@ class MLS {
     long  getMovement();
     float getDistance(float lat1, float long1, float lat2, float long2);
     int   getBearing(float lat1, float long1, float lat2, float long2);
-    const char *getCardinal(float course);
+    const char* getCardinal(int course);
     geo_t current;
     geo_t previous;
     float distance;
